Check for database tables before creating schema
When a table already exists, create_table? silently does nothing.
So after changing schema.rb and running 'rake db:create' the database
might still have the outdated schema and not the expected one.

Instead, output a warning and exit when tables already exist, providing
immediate feedback that the database might not be in the expected state.